+++ v0.4/plugins/login-impl/src/components/LoginForm.svelte
@@ -14,7 +14,13 @@
 -->
 
 <script lang="ts">
+  import { getContext } from 'svelte'
+  import { Platform } from '@anticrm/plugin'
+
   import Form from './Form.svelte'
+  import { doLogin } from '../utils'
+
+  const platform = getContext('platform') as Platform
 
   const fields = [
     { name: 'username', i18n: 'Email' },
@@ -26,9 +32,16 @@
     { name: 'workspace', i18n: 'Workspace' }
   ]
 
+  let object = {
+    workspace: '',
+    username: '',
+    password: '',
+  }
+
   const action = { 
     i18n: 'Login',
     func: async () => { 
+      // const [loginStatus, result] = await doLogin(platform, object.username, object.password, object.workspace)
       console.log('loging in...') 
       return new Promise<void>((resolve, reject) => {
         setTimeout(() => { resolve() }, 1000)
@@ -36,11 +49,7 @@
     }
   }
 
-  let object = {
-    workspace: '',
-    username: '',
-    password: '',
-  }
+
 </script>
 
 <Form caption="Log In" {fields} {object} {action}/>
